feat(FOROME-950): add css module
