+++ src/pages/filter/ui/query-builder/ui/all-not-mods.tsx
@@ -32,7 +32,7 @@ export const AllNotMods = observer(
         {isAllModeAvailable && (
           <>
             <Checkbox
-              id={Math.random()}
+              id={'all-mode-checkbox'}
               className="flex items-center"
               checked={isAllModeChecked ?? false}
               disabled={isAllModeDisabled}
@@ -46,7 +46,7 @@ export const AllNotMods = observer(
         )}
 
         <Checkbox
-          id={Math.random()}
+          id={'not-mode-checkbox'}
           className="flex items-center"
           checked={isNotModeChecked ?? false}
           disabled={isNotModeDisabled}
