+++ src/ui/check-input/check-input.module.css
@@ -1,0 +1,50 @@
+.checkbox {
+    display: none;
+  }
+
+.checkmark {
+    position: relative;
+    display: block;
+    width: 16px;
+    height: 16px;
+    border: 1px solid theme('colors.grey.blue');
+    border-radius: 0.125rem;
+    transition: all 0.2s;
+    cursor: pointer;
+}
+
+.label {
+    margin-left: 8px;
+    cursor: pointer;
+}
+
+.checkmark::after {
+    content: '';
+    position: absolute;
+    top: 45%;
+    left: 50%;
+    width: 5px;
+    height: 9px;
+    transform: translate(-50%, -50%) rotateZ(40deg);
+    border-right: 1px solid transparent;
+    border-bottom: 1px solid transparent;
+}
+
+.checkbox:checked + .checkmark {
+    background-color: theme('colors.blue.bright');
+    border: 1px solid theme('colors.blue.bright');
+}
+
+.checkbox:disabled:checked + .checkmark {
+    background-color: theme('colors.grey.disabled');
+    border: 1px solid theme('colors.grey.disabled');
+}
+
+.checkbox:checked + .checkmark::after {
+    border-right: 1px solid theme('colors.white');
+    border-bottom: 1px solid theme('colors.white');
+}
+
+.checkbox:checked ~ .label {
+    font-weight: 700;
+}
