+++ src/ui/checkbox/checkbox.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react'
 
-import { CheckInput, ICheckInputProps } from '@ui/check-input'
+import { CheckInput, ICheckInputProps } from '@ui/check-input/check-input'
 
 export const Checkbox: FC<Omit<ICheckInputProps, 'type'>> = props => (
   <CheckInput type="checkbox" {...props} />
