+++ src/ui/radio.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react'
 
-import { CheckInput, ICheckInputProps } from '@ui/check-input'
+import { CheckInput, ICheckInputProps } from '@ui/check-input/check-input'
 
 export const Radio: FC<Omit<ICheckInputProps, 'type'>> = props => (
   <CheckInput type="radio" {...props} />
